requirements.txt
astroid==2.12.13
black==22.12.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dill==0.3.6
fonttools==4.39.3
importlib-resources==5.12.0
isort==5.11.4
kiwisolver==1.4.4
lazy-object-proxy==1.8.0
matplotlib==3.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.24.1
packaging==23.0
pathspec==0.10.3
Pillow==9.5.0
platformdirs==2.6.0
pylint==2.15.9
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
scipy==1.10.1
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.65.0
typing_extensions==4.4.0
wrapt==1.14.1
zipp==3.15.0
